Move tabs into their own directory
