+++ src/examples/tabs-example.js
@@ -1,10 +1,7 @@
 import React from 'react';
 
 import Example from './example.js';
-import Tabs from '../tabs.js';
-import TabList from '../tab-list.js';
-import Tab from '../tab.js';
-import TabPanels from '../tab-panels.js';
+import { Tabs, TabList, Tab, TabPanels } from '../tabs';
 
 export default class TabsExample extends React.Component {
   constructor() {
