+++ src/tabs/index.js
@@ -1,0 +1,4 @@
+export { TabList  as TabList} from './tab-list.js';
+export { TabPanels as TabPanels } from './tab-panels.js';
+export { Tab as Tab } from './tab.js';
+export { Tabs as Tabs } from './tabs.js';
